fix warnings

+++ VTK/Common/vtkOrderedTriangulator.cxx
@@ -981,7 +981,7 @@ void vtkOTMesh::DumpInsertionCavity(double x[3])
   cout << "DATASET POLYDATA\n";
 
   //write out points
-  int numFaces = this->CavityFaces.size();
+  int numFaces = (int)this->CavityFaces.size();
   cout << "POINTS " << 3*numFaces+1 << " float\n";
   
   for (fptr=this->CavityFaces.begin();
